feat: ensure RemoteAuthentication remove favorites that belong to it

+++ src/modules/favorites/services/RemoveFavoriteService.ts
@@ -1,5 +1,6 @@
 import { injectable, inject } from 'tsyringe';
 import IFavoritesRepository from '../repositories/IFavoritesRepository';
+import AppError from '../../../shared/errors/AppError';
 
 @injectable()
 class RemoveFavoriteService {
@@ -8,7 +9,16 @@ class RemoveFavoriteService {
     private favoriteRepository: IFavoritesRepository,
   ) {}
 
-  async execute(favoriteId: string): Promise<void> {
+  async execute(user_id: string, favoriteId: string): Promise<void> {
+    const favorite = await this.favoriteRepository.findById(
+      user_id,
+      favoriteId,
+    );
+
+    if (!favorite) {
+      throw new AppError('This product does not belong to you!', 403);
+    }
+
     await this.favoriteRepository.remove(favoriteId);
   }
 }
